[spring-cloud-gateway] refactor : 일부 모듈 및 내부 컴포넌트 이름 변경
