--- playground/spring-cloud-gateway/spring-cloud-gateway-reactive-server-1/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "spring-cloud-gateway-reactive-server-1"
