--- playground/spring-cloud-gateway/spring-cloud-gateway-reactive-server-1/src/main/kotlin/com/banjjoknim/springcloudgatewayreactiveserver1/SpringCloudGatewayReactiveServer1Application.kt
@@ -1,11 +1,0 @@
-package com.banjjoknim.springcloudgatewayreactiveserver1
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class SpringCloudGatewayReactiveServer1Application
-
-fun main(args: Array<String>) {
-    runApplication<SpringCloudGatewayReactiveServer1Application>(*args)
-}
