--- playground/spring-cloud-gateway/spring-cloud-gateway-reactive-server-2/src/main/resources/application.yaml
@@ -1,6 +1,0 @@
-ktor:
-  application:
-    modules:
-      - com.banjjoknim.ApplicationKt.module
-  deployment:
-    port: 8020
