+++ playground/spring-cloud-gateway/spring-cloud-gateway-server-1/.gitattributes
