+++ playground/spring-cloud-gateway/spring-cloud-gateway-server-1/src/main/kotlin/com/banjjoknim/springcloudgatewayserver1/SpringCloudGatewayServer1Application.kt
@@ -1,0 +1,11 @@
+package com.banjjoknim.springcloudgatewayserver1
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class SpringCloudGatewayServer1Application
+
+fun main(args: Array<String>) {
+    runApplication<SpringCloudGatewayServer1Application>(*args)
+}
