+++ playground/spring-cloud-gateway/spring-cloud-gateway-server-2/settings.gradle.kts
@@ -1,0 +1,1 @@
+rootProject.name = "spring-cloud-gateway-server-2"
