+++ playground/spring-cloud-gateway/spring-cloud-gateway-server-2/src/main/resources/application.yaml
@@ -1,0 +1,6 @@
+ktor:
+  application:
+    modules:
+      - com.banjjoknim.SpringCloudGatewayServer2ApplicationKt.module
+  deployment:
+    port: 8020
