+++ playground/spring-cloud-gateway/spring-cloud-gateway-server-2/src/main/resources/logback.xml
